# Conflicts:
#	mge_v18_student_version/_vs2015/debug_static_build/AbstractGame.obj
#	mge_v18_student_version/_vs2015/debug_static_build/GridComponent.obj
#	mge_v18_student_version/_vs2015/debug_static_build/KeyMoveComponent.obj
#	mge_v18_student_version/_vs2015/debug_static_build/MainGame.obj
#	mge_v18_student_version/_vs2015/debug_static_build/MainWorld.obj
#	mge_v18_student_version/_vs2015/debug_static_build/ShootingComponent.obj
#	mge_v18_student_version/_vs2015/debug_static_build/TileComponent.obj
#	mge_v18_student_version/_vs2015/debug_static_build/mge.tlog/CL.command.1.tlog
#	mge_v18_student_version/_vs2015/debug_static_build/mge.tlog/CL.read.1.tlog
#	mge_v18_student_version/_vs2015/debug_static_build/mge.tlog/CL.write.1.tlog
#	mge_v18_student_version/_vs2015/debug_static_build/mge.tlog/link.read.1.tlog
#	mge_v18_student_version/_vs2015/debug_static_build/vc141.idb
#	mge_v18_student_version/_vs2015/debug_static_exe/mge.exe
#	mge_v18_student_version/src/game/components/ShootingComponent.cpp

+++ mge_v18_student_version/src/game/components/TileComponent.cpp
@@ -15,12 +15,12 @@ void TileComponent::Parse(rapidxml::xml_node<>* compNode)
 {
 }
 
-void TileComponent::SetGridPos(glm::vec2 gridPos)
+void TileComponent::SetGridPos(glm::ivec2 gridPos)
 {
 	m_gridPos = gridPos;
 }
 
-glm::vec2 TileComponent::GridPos()
+glm::ivec2 TileComponent::GridPos()
 {
 	return m_gridPos;
 }
